pyproject.toml
[tool.poetry]
name = "saleor"
version = "2.9.0"
description = "A modular, high performance e-commerce storefront built with GraphQL, Django, and ReactJS."
authors = ["Mirumee Software <hello@mirumee.com>"]
license = "BSD-3-Clause"

readme = "README.md"

homepage = "https://getsaleor.com/"
repository = "https://github.com/mirumee/saleor"
documentation = "https://docs.getsaleor.com/en/latest/"

[tool.poetry.dependencies]
python = "~3.8"
babel = "~2.7"
bleach = "~2"
boto3 = "^1.9"
braintree = "~3.49.0"
celery = { version = "^4.3.0", extras = ["redis"] }
dj-database-url = "^0"
dj-email-url = "^0"
django = "~2.2"
django-babel = "^0"
django-countries = "^5.3"
django-filter = "^2.2"
django-graphql-jwt = "^0.3.0" # https://github.com/mirumee/saleor/issues/4652
django-measurement = "^3.0"
django-mptt = "^0"
django-phonenumber-field = "^2.4"
django-prices = "^2.1"
django-prices-openexchangerates = "^1.0.1"
django-prices-vatlayer = "^1.0.2"
django-silk = "~2.0"
django-storages = { version = "^1.7.1", extras = [ "google" ] }
django-templated-email = "^2.3.0"
django-versatileimagefield = "^1.11"
draftjs-sanitizer = "^1.0.0"
faker = "^2.0"
freezegun = "^0"
google-cloud-storage = "^1.18.0"
google-i18n-address = "^2.3.5"
google-measurement-protocol = "^1.0"
graphene-django = "^2.8.0"
graphene-django-optimizer = "^0.6"
graphene-federation = "^0.0.3"
html-to-draftjs = "^1.0.1"
markdown = "^3.1.1"
maxminddb = "^1.4.1"
maxminddb-geolite2 = "^2018.701"
opentracing = "^2.3.0"
phonenumberslite = "^8.10.16"
prices = "^1.0"
psycopg2-binary = "^2.8.3"
purl = "^1.5"
python-magic-bin = {version = "^0.4.14", platform = "win32"}
razorpay = "^1.2"
requests = "^2.22"
sentry-sdk = "^0"
stripe = "^2.33.2"
text-unidecode = "^1.2"
tqdm = "^4.36"
uwsgi = {version = "^2.0.8", platform = "!=win32"}
weasyprint = ">=48"
oauthlib = "^3.0"

[tool.poetry.dev-dependencies]
black = "19.10b0"
codecov = "^2.0.15"
coverage = "^4.5.4"
django-debug-toolbar = "^2.0"
django-debug-toolbar-request-history = "^0"
django-graphiql-debug-toolbar = "^0.1.4"
django-extensions = "^2.2.1"
flake8 = "^3.7.8"
isort = "^4.3.21"
pre-commit = "^1.18"
pycodestyle = "^2.5"
pydocstyle = "^4.0"
pylint = "^2.3.1"
pylint-celery = "^0"
pylint-django = "^2.0.11"
pylint-plugin-utils = "^0"
pytest = "^5.0.1"
pytest-cov = "^2.7.1"
pytest-django = "^3.5.1"
pytest-django-queries = "~1.1"
pytest-mock = "^1.10.4"
pytest-vcr = "^1.0.2"
pytest-xdist = "^1.29"
tox = "^3.13.2"
transifex-client = "^0"
django-stubs = "^1.2"
mypy = "^0.740.0"

[tool.black]
target_version = ['py35', 'py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(\.git/
  |\.eggs
  |\.hg
  |__pycache__
  |\.cache
  |\.ipynb_checkpoints
  |\.mypy_cache
  |\.pytest_cache
  |\.tox
  |\.venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |infrastructure
  |templates
  |locale
)/
'''
